modify: 채팅/친구 탭과 컴포넌트 연결
