+++ apps/front/src/components/Main/Tab/index.tsx
@@ -1,28 +1,24 @@
 import { useState } from "react";
 import styles from "./index.module.scss";
 
-function Tab() {
-  const CHAT = "chat";
-  const FRIEND = "friend";
-  const tabs = [
-    { id: CHAT, label: "채팅방" },
-    { id: FRIEND, label: "친구" },
-  ];
-  const [selectedTab, setSelectedTab] = useState<string>(CHAT);
-
-  const onClickTabItem = (e: React.MouseEvent<HTMLInputElement>) => {
-    setSelectedTab(e.currentTarget.id);
-  };
-
+interface TabProps {
+  tabs: {
+    id: string;
+    label: string;
+  }[];
+  currentTab: string;
+  handleTab: (e: React.MouseEvent<HTMLInputElement>) => void;
+}
+function Tab({ tabs, currentTab, handleTab }: TabProps) {
   return (
     <div className={styles.container}>
       {tabs.map((item) => (
         <div
           key={item.id}
           id={item.id}
-          onClick={onClickTabItem}
+          onClick={handleTab}
           className={`${styles.item} ${
-            selectedTab === item.id ? styles.selected : ""
+            currentTab === item.id ? styles.selected : ""
           }`}
         >
           {item.label}
