+++ apps/front/src/components/Main/index.tsx
@@ -1,3 +1,4 @@
+import { useState } from "react";
 import Chat from "components/Chat";
 import DefaultHeader from "components/DefaultHeader";
 import Friends from "components/Friends";
@@ -5,14 +6,24 @@ import styles from "./index.module.scss";
 import Tab from "./Tab";
 
 function Main() {
+  const CHAT = "chat";
+  const FRIEND = "friend";
+  const tabs = [
+    { id: CHAT, label: "채팅방" },
+    { id: FRIEND, label: "친구" },
+  ];
+  const [currentTab, setCurrentTab] = useState<string>(CHAT);
+  const handleTab = (e: React.MouseEvent<HTMLInputElement>) => {
+    setCurrentTab(e.currentTarget.id);
+  };
   return (
     <div className={styles.container}>
       <DefaultHeader />
       <div className={styles.gap} />
-      <Tab />
+      <Tab tabs={tabs} currentTab={currentTab} handleTab={handleTab} />
       <div className={styles.gap} />
-      {/* <Chat /> */}
-      <Friends />
+      {currentTab === CHAT && <Chat />}
+      {currentTab === FRIEND && <Friends />}
     </div>
   );
 }
